pyproject.toml
[project]
name = "donation-ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aio-pika>=9.5.8",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
